Avoid loading SpinnakerNET during deserialization
Move the camera image event handler, and the frame converter it uses,
out of SpinnakerCapture into a top-level ImageEventListener class.
SpinnakerCapture no longer declares a nested type whose base type is in
SpinnakerNET, so building the XmlSerializer for a workflow no longer
forces SpinnakerNET to load, and the workflow opens and runs without
the SDK installed.
